moved controls to master track
